--- vavedem-interfaces/src/main/java/ro/vavedem/interfaces/database/IRepository.java
@@ -1,10 +1,0 @@
-package ro.vavedem.interfaces.database;
-
-/**
- *
- * @author CoruptiaUcide
- */
-public interface IRepository {
-
-
-}
