--- vavedem-persistence/src/main/java/ro/vavedem/persistence/repository/LocalitateCRUDRepository.java
@@ -1,16 +1,0 @@
-package ro.vavedem.persistence.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import ro.vavedem.persistence.entities.Localitate;
-
-import java.util.List;
-
-public interface LocalitateCRUDRepository extends CrudRepository<Localitate, Long> {
-
-    Localitate findOne(Long id);
-
-    List<Localitate> findAll();
-
-    Localitate save(Localitate primarie);
-
-}
