+++ vavedem-persistence/src/main/java/ro/vavedem/persistence/repository/LocalitateRepository.java
@@ -1,14 +1,16 @@
-package ro.vavedem.persistence.service;
+package ro.vavedem.persistence.repository;
 
+import org.springframework.data.repository.CrudRepository;
 import ro.vavedem.persistence.entities.Localitate;
 
 import java.util.List;
 
-public interface LocalityService {
+public interface LocalitateRepository extends CrudRepository<Localitate, Long> {
 
     Localitate findOne(Long id);
 
     List<Localitate> findAll();
 
     Localitate save(Localitate primarie);
+
 }
