--- vavedem-persistence/src/main/java/ro/vavedem/persistence/repository/PrimarieCRUDRepository.java
@@ -1,20 +1,0 @@
-package ro.vavedem.persistence.repository;
-
-import org.springframework.data.repository.CrudRepository;
-import ro.vavedem.persistence.entities.Primarie;
-
-import java.util.List;
-
-public interface PrimarieCRUDRepository extends CrudRepository<Primarie, Long> {
-
-    Primarie findOne(Long id);
-
-    List<Primarie> findAll();
-
-    Primarie save(Primarie primarie);
-
-    void delete(Primarie primarie);
-
-    List<Primarie> findByNume(String nume);
-
-}
