+++ vavedem-services/src/main/java/ro/vavedem/services/LocalitateService.java
@@ -9,22 +9,22 @@ import ro.vavedem.exceptions.VaVedemPersistenceException;
 import ro.vavedem.interfaces.database.Service;
 import ro.vavedem.models.LocalitateModel;
 import ro.vavedem.persistence.entities.Localitate;
-import ro.vavedem.persistence.service.LocalityService;
+import ro.vavedem.persistence.repository.LocalitateRepository;
 import ro.vavedem.services.util.LocalitateServUtil;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @org.springframework.stereotype.Service
-public class LocalitateConversionService implements Service<LocalitateModel> {
+public class LocalitateService implements Service<LocalitateModel> {
 
-    private static final Logger logger = Logger.getLogger(LocalitateConversionService.class);
+    private static final Logger logger = Logger.getLogger(LocalitateService.class);
 
     @Autowired
-    private LocalityService localityService;
+    private LocalitateRepository repository;
 
     public LocalitateModel findOne(Long id) throws VaVedemApiException {
-        final Localitate entity = localityService.findOne(id);
+        final Localitate entity = repository.findOne(id);
 
         if (null == entity) {
             throw new VaVedemNotFoundException("Not found any record with id: " + id);
@@ -35,7 +35,7 @@ public class LocalitateConversionService implements Service<LocalitateModel> {
 
     public List<LocalitateModel> findAll() throws VaVedemApiException {
         final List<LocalitateModel> models = new ArrayList<>();
-        final List<Localitate> entities = localityService.findAll();
+        final List<Localitate> entities = repository.findAll();
 
         for (Localitate e : entities) {
             models.add(LocalitateServUtil.convertToModel(e));
@@ -47,7 +47,7 @@ public class LocalitateConversionService implements Service<LocalitateModel> {
 
     public LocalitateModel save(final LocalitateModel model) throws VaVedemApiException {
         final Localitate p = LocalitateServUtil.convertToEntity(model);
-        final Localitate saved = localityService.save(p);
+        final Localitate saved = repository.save(p);
 
         if (null == saved) {
             throw new VaVedemPersistenceException("Fail to save the entity.");
